move DirectSumTraversal(Interface) to traversals directory
